refactor: consolidate scraper and utility logic into common and raw
