+++ src/main.py
@@ -7,8 +7,8 @@ from pathlib import Path
 from zoneinfo import ZoneInfo
 
 from catalog.manage_iceberg import get_catalog, provision_table
+from common.jepx_common import resolve_target_at
 from common.storage_client import RustFSClient
-from jepx.common import resolve_target_at
 from orchestration.jepx_pipeline import run_jepx_orchestrated_pipeline
 from pipeline.bronze.ingest_jepx import (
     ingest_jepx_spot_summary,
@@ -16,13 +16,15 @@ from pipeline.bronze.ingest_jepx import (
 )
 from pipeline.bronze.ingest_occto import ingest_occto_unit_generation
 from pipeline.bronze.migrate_occto_data import migrate_occto_data
-from pipeline.raw.jepx_to_rustfs import scrape_jepx_to_rustfs
-from scraper.module.jepx import JEPXSpotSummaryScraper
-from scraper.module.occto import (
+from pipeline.raw.source_to_raw_jepx import (
+    JEPXSpotSummaryScraper,
+    scrape_jepx_to_rustfs,
+)
+from pipeline.raw.source_to_raw_occto import (
     OCCTOUnitGenerationConfig,
     OCCTOUnitGenerationScraper,
+    scrape_occto_to_rustfs,
 )
-from scraper.occto_to_rustfs import scrape_occto_to_rustfs
 from setup.rustfs_bucket_setup import BucketPlan, apply_bucket_plans
 
 logging.basicConfig(
