+++ src/orchestration/jepx_pipeline.py
@@ -20,11 +20,13 @@ import duckdb
 import polars as pl
 
 from catalog.manage_iceberg import get_catalog, provision_table
+from common.jepx_common import resolve_target_at
 from common.storage_client import RustFSClient
-from jepx.common import resolve_target_at
 from pipeline.bronze.ingest_jepx import ingest_jepx_spot_summary
-from pipeline.raw.jepx_to_rustfs import scrape_jepx_to_rustfs
-from scraper.module.jepx import JEPXSpotSummaryScraper
+from pipeline.raw.source_to_raw_jepx import (
+    JEPXSpotSummaryScraper,
+    scrape_jepx_to_rustfs,
+)
 
 logger = logging.getLogger(__name__)
