+++ src/pipeline/bronze/ingest_jepx.py
@@ -10,13 +10,13 @@ import polars as pl
 sys.path.append(str(Path(__file__).resolve().parent.parent.parent))
 
 from catalog.manage_iceberg import get_catalog
-from common.storage_client import RustFSClient
-from jepx.common import (
+from common.jepx_common import (
     resolve_spot_summary_file_name,
     resolve_spot_summary_object_key,
     resolve_target_at,
 )
-from utility.pipeline_utilities import add_metadata, build_schema_exprs
+from common.pipeline_utilities import add_metadata, build_schema_exprs
+from common.storage_client import RustFSClient
 
 # ログ設定
 logging.basicConfig(
