+++ src/pipeline/raw/source_to_raw_occto.py
@@ -1,9 +1,12 @@
+"""OCCTO source-to-raw scraping and upload workflow."""
+
 from __future__ import annotations
 
 from dataclasses import dataclass
 from datetime import UTC, date, datetime
 
-from scraper.module.base import BaseHttpScraper, RequestSpec
+from common.http_scraper import BaseHttpScraper, RequestSpec
+from common.storage_client import RustFSClient
 
 
 @dataclass(frozen=True)
@@ -83,3 +86,37 @@ class OCCTOUnitGenerationScraper(BaseHttpScraper):
             file_name=file_name,
             body=body,
         )
+
+
+@dataclass(frozen=True)
+class OCCTOScrapeUploadResult:
+    """Upload result metadata for an OCCTO scraped CSV."""
+
+    bucket_name: str
+    object_key: str
+    file_name: str
+    size_bytes: int
+
+
+def scrape_occto_to_rustfs(
+    storage_client: RustFSClient,
+    scraper: OCCTOUnitGenerationScraper,
+    bucket_name: str,
+    target_at: date | datetime,
+) -> OCCTOScrapeUploadResult:
+    """Scrape OCCTO CSV for a target date and upload it to RustFS raw layer."""
+    scraped = scraper.scrape(target_at)
+
+    storage_client.upload_bytes(
+        bucket_name=bucket_name,
+        object_name=scraped.object_key,
+        body=scraped.body,
+        content_type=scraped.content_type,
+    )
+
+    return OCCTOScrapeUploadResult(
+        bucket_name=bucket_name,
+        object_key=scraped.object_key,
+        file_name=scraped.file_name,
+        size_bytes=len(scraped.body),
+    )
